# public/openapi.generated.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Life@USTC API",
    "version": "1.0.0",
    "description": "OpenAPI document generated from Next.js routes by next-openapi-gen"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local server"
    }
  ],
  "paths": {
...
    "/api/sections": {
      "get": {
        "operationId": "get-api-sections",
        "summary": "List sections with filters and pagination.",
        "description": "",
        "tags": ["Api"],
        "parameters": [
          {
            "in": "query",
            "name": "courseId",
            "schema": {},
            "required": false
          },
          {
            "in": "query",
            "name": "semesterId",
            "schema": {},
            "required": false
          },
          {
            "in": "query",
            "name": "campusId",
            "schema": {},
            "required": false
          },
          {
            "in": "query",
            "name": "departmentId",
            "schema": {},
            "required": false
          },
          {
            "in": "query",
            "name": "teacherId",
            "schema": {},
            "required": false
          },
          {
            "in": "query",
            "name": "ids",
            "schema": {
              "type": "string"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "page",
            "schema": {},
            "required": false
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {},
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/paginatedSectionResponseSchema"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/teachers": {
      "get": {
        "operationId": "get-api-teachers",
        "summary": "List teachers with department/search filters.",
        "description": "",
        "tags": ["Api"],
        "parameters": [
          {
            "in": "query",
            "name": "departmentId",
            "schema": {},
            "required": false
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "page",
            "schema": {},
            "required": false
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {},
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/paginatedTeacherResponseSchema"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/homeworks/{id}": {
      "patch": {
        "operationId": "patch-api-homeworks-{id}",
        "summary": "Update one homework.",
        "description": "",
        "tags": ["Api"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "123"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/homeworkUpdateRequestSchema"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/successResponseSchema"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "delete-api-homeworks-{id}",
        "summary": "Soft delete one homework.",
        "description": "",
        "tags": ["Api"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "123"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
    "/api/uploads/{id}/download": {
      "get": {
        "operationId": "get-api-uploads-{id}-download",
        "summary": "Redirect to signed URL for one upload.",
        "description": "",
        "tags": ["Api"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "123"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "successResponseSchema": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          }
        },
        "required": ["success"]
      },
...
      "homeworkUpdateRequestSchema": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "publishedAt": {
            "type": "string",
            "nullable": true
          },
          "submissionStartAt": {
            "type": "string",
            "nullable": true
          },
          "submissionDueAt": {
            "type": "string",
            "nullable": true
          },
          "isMajor": {
            "type": "boolean"
          },
          "requiresTeam": {
            "type": "boolean"
          }
        }
      },
...
      "paginatedSectionResponseSchema": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/sectionListSchema"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/pageMetaSchema"
          }
        },
        "required": ["data", "pagination"]
      },
      "semesterSchema": {
        "type": "object",
        "properties": {
          "startDate": {
            "allOf": [
              {
                "$ref": "#/components/schemas/dateTimeSchema"
              }
            ]
          },
          "endDate": {
            "allOf": [
              {
                "$ref": "#/components/schemas/dateTimeSchema"
              }
            ]
          }
        },
        "required": ["startDate", "endDate"]
      },
...
      "paginatedTeacherResponseSchema": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/teacherListSchema"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/pageMetaSchema"
          }
        },
        "required": ["data", "pagination"]
      },
...
      "dateTimeSchema": {
        "type": "string"
      },
      "pageMetaSchema": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer"
          },
          "pageSize": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          },
          "totalPages": {
            "type": "integer"
          }
        },
        "required": ["page", "pageSize", "total", "totalPages"]
      },
...
      "courseSchema": {
        "type": "object",
        "properties": {
          "category": {
            "type": "object"
          },
          "classType": {
            "type": "object"
          },
          "classify": {
            "type": "object"
          },
          "educationLevel": {
            "type": "object"
          },
          "gradation": {
            "type": "object"
          },
          "type": {
            "type": "object"
          }
        },
        "required": [
          "category",
          "classType",
          "classify",
          "educationLevel",
          "gradation",
          "type"
        ]
      },
      "departmentSchema": {
        "type": "object"
      },
      "teacherSchema": {
        "type": "object"
      },
...
      "campusSchema": {
        "type": "object"
      },
      "teachLanguageSchema": {
        "type": "object"
      },
      "examModeSchema": {
        "type": "object"
      },
...
      "teacherTitleSchema": {
        "type": "object"
      },
...
      "adminClassSchema": {
        "type": "object"
      },
      "sectionListSchema": {
        "type": "object",
        "properties": {
          "course": {
            "$ref": "#/components/schemas/courseSchema"
          },
          "semester": {
            "allOf": [
              {
                "$ref": "#/components/schemas/semesterSchema"
              }
            ]
          },
          "campus": {
            "allOf": [
              {
                "$ref": "#/components/schemas/campusSchema"
              }
            ]
          },
          "openDepartment": {
            "allOf": [
              {
                "$ref": "#/components/schemas/departmentSchema"
              }
            ]
          },
          "examMode": {
            "allOf": [
              {
                "$ref": "#/components/schemas/examModeSchema"
              }
            ]
          },
          "teachLanguage": {
            "allOf": [
              {
                "$ref": "#/components/schemas/teachLanguageSchema"
              }
            ]
          },
          "teachers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/teacherSchema"
            }
          },
          "adminClasses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/adminClassSchema"
            }
          }
        },
        "required": [
          "course",
          "semester",
          "campus",
          "openDepartment",
          "examMode",
          "teachLanguage",
          "teachers",
          "adminClasses"
        ]
      },
      "teacherListSchema": {
        "type": "object",
        "properties": {
          "department": {
            "allOf": [
              {
                "$ref": "#/components/schemas/departmentSchema"
              }
            ]
          },
          "teacherTitle": {
            "allOf": [
              {
                "$ref": "#/components/schemas/teacherTitleSchema"
              }
            ]
          },
          "_count": {
            "type": "object",
            "properties": {
              "sections": {
                "type": "integer"
              }
            },
            "required": ["sections"]
          }
        },
        "required": ["department", "teacherTitle", "_count"]
      }
    },
...
  }
}